moved hook into helpers
