+++ src/helpers/setBeforeUnload.ts
